replace tests to correct file
