+++ tests/opentelemetry-docker-tests/tests/postgres/test_aiopg_functional.py
@@ -42,6 +42,10 @@ class TestFunctionalAiopgConnect(TestBase):
         cls._cursor = None
         cls._tracer = cls.tracer_provider.get_tracer(__name__)
         AiopgInstrumentor().instrument(tracer_provider=cls.tracer_provider)
+        cls._dsn = (
+            f"dbname='{POSTGRES_DB_NAME}' user='{POSTGRES_USER}' password='{POSTGRES_PASSWORD}'"
+            f" host='{POSTGRES_HOST}' port='{POSTGRES_PORT}'"
+        )
         cls._connection = async_call(
             aiopg.connect(
                 dbname=POSTGRES_DB_NAME,
@@ -185,3 +189,19 @@ class TestFunctionalAiopgCreatePool(TestBase):
         ):
             async_call(self._cursor.callproc("test", ()))
             self.validate_spans("test")
+
+    def test_instrumented_pool_with_multiple_acquires(self, *_, **__):
+        async def double_asquire():
+            pool = await aiopg.create_pool(dsn=self._dsn)
+            async with pool.acquire() as conn:
+                async with conn.cursor() as cursor:
+                    query = "SELECT 1"
+                    await cursor.execute(query)
+            async with pool.acquire() as conn:
+                async with conn.cursor() as cursor:
+                    query = "SELECT 1"
+                    await cursor.execute(query)
+
+        async_call(double_asquire())
+        spans = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans), 2)
